Refactor MIDI helpers
