+++ midigen.py
@@ -218,123 +218,7 @@ from midi2audio import FluidSynth
 from collections import defaultdict
 import os # Import os for file operations
 
-# -------- MIDI Parser --------
-def midi_to_poly_dataset(midi_file):
-    try:
-        if hasattr(midi_file, "read"):
-            mid = mido.MidiFile(file=midi_file)
-        else:
-            mid = mido.MidiFile(midi_file)
-        
-        all_notes_for_ai = [] # List to store all notes from all tracks
-        
-        for i, mtrack in enumerate(mid.tracks):
-            active_notes = {}
-            abs_tick = 0
-            
-            for msg in mtrack:
-                abs_tick += msg.time
-                if msg.type == 'note_on' and msg.velocity > 0:
-                    active_notes[(msg.note, msg.channel)] = abs_tick
-                elif msg.type == 'note_off' or (msg.type == 'note_on' and msg.velocity == 0):
-                    if (msg.note, msg.channel) in active_notes:
-                        start_tick = active_notes.pop((msg.note, msg.channel))
-                        duration = abs_tick - start_tick
-                        if duration > 0:
-                            all_notes_for_ai.append([
-                                start_tick, # Store absolute tick for sorting
-                                msg.note,
-                                msg.velocity,
-                                duration,
-                                i # track index
-                            ])
-
-            # Handle notes that are still active at the end of the track
-            for (note, channel), start_tick in active_notes.items():
-                duration = abs_tick - start_tick # Assume duration until end of track
-                if duration > 0:
-                    all_notes_for_ai.append([
-                        start_tick, # Store absolute tick for sorting
-                        note,
-                        64, # Default velocity for sustained notes
-                        duration,
-                        i # track index
-                    ])
-
-        # Sort all notes by absolute tick for interleaved sequence
-        all_notes_for_ai.sort(key=lambda x: x[0])
-
-        # Convert absolute ticks to delta ticks for the AI model
-        final_dataset = []
-        last_abs_tick = 0
-        for note_event in all_notes_for_ai:
-            abs_tick, note, velocity, duration, track_idx = note_event
-            delta_tick = abs_tick - last_abs_tick
-            final_dataset.append([delta_tick, note, velocity, duration, track_idx])
-            last_abs_tick = abs_tick
-        
-        return np.array(final_dataset) if final_dataset else np.array([])
-    except Exception as e:
-        st.error(f"Erro ao processar arquivo MIDI: {e}")
-        return np.array([])
-
-def notes_to_midi_file(notes, tempo=120):
-    mid = mido.MidiFile()
-    
-    # Create tracks in mido.MidiFile
-    # Determine the maximum track index to create enough tracks
-    max_track_idx = 0
-    if notes:
-        for track_data in notes:
-            if track_data.size > 0:
-                max_track_idx = max(max_track_idx, int(track_data[:, 4].max()))
-    
-    for i in range(max_track_idx + 1):
-        mid.add_track()
-
-    # Add notes to mido tracks
-    for track_idx, track_data in enumerate(notes):
-        # We need to convert delta_tick to absolute_tick for sorting and for note_on/off messages.
-        cumulative_tick = 0
-        absolute_notes_data = []
-        for note_event in track_data:
-            delta_tick, note, velocity, duration, _ = note_event
-            cumulative_tick += delta_tick # Accumulate delta_tick
-            absolute_notes_data.append([cumulative_tick, note, velocity, duration, track_idx])
-
-        # Now sort by absolute_tick
-        if len(absolute_notes_data) > 0:
-            sorted_track_data = np.array(absolute_notes_data)[np.argsort(np.array(absolute_notes_data)[:, 0])]
-        else:
-            sorted_track_data = []
-        
-        # Create a list of (tick, message) pairs for this track
-        track_messages = []
-        for note_event in sorted_track_data:
-            start_tick, note, velocity, duration, _ = note_event
-            
-            # Note On message
-            track_messages.append((start_tick, mido.Message("note_on", note=int(note), velocity=int(velocity))))
-            
-            # Note Off message
-            end_tick = start_tick + duration
-            track_messages.append((end_tick, mido.Message("note_off", note=int(note), velocity=int(velocity))))
-
-        # Sort all messages by tick and then convert to delta time
-        track_messages.sort(key=lambda x: x[0])
-        
-        last_tick = 0
-        for tick, msg in track_messages:
-            delta_time = int(tick - last_tick)
-            msg.time = delta_time # Set delta time for the message
-            mid.tracks[track_idx].append(msg)
-            last_tick = tick
-
-    # Add tempo message to the first track (or a dedicated tempo track)
-    if mid.tracks:
-        mid.tracks[0].insert(0, mido.MetaMessage("set_tempo", tempo=mido.bpm2tempo(tempo)))
-
-    return mid
+from midi_utils import midi_to_poly_dataset, notes_to_midi_file, export_midi_file
 
 # -------- IA Polifônica --------
 class PolyTransformer(nn.Module):
